Implemented a Starting GUI
Note: Can not currently create new Task and have task displayed by window as waiting for that code. In addition, code dealing with controller and update is not implemented either. This is just to give a launch pad for others to work off of. Finally code right now only deals with 1 list. In future, will have to have a way to view multiple lists.