Saving and Loading Implemented
Note: A lot of things changed while I was implementating it in order to 1) Make Junit testing easier later (having the model be passed to the controller, this is so the Junit can create and access the model passed to the controller) and 2) make MVC seperation better and 3) make saving easier. Before, controller was also doing some model responsibilities as it was also going to be in charge of creating more ToDoLists and storing them when that should probably be the model's job.