Multiple List Functionality Implemented
Can create new lists which are added to the end of the list. Can delete
the current lists (only if there is more than 1 list though). Lists can
be renamed. The saving works with the multiple lists. The addObserver
methods were streamlined, so that the view only has to use that method
once. Update method had to have the object being passed to it changed
from ToDoTask to ToDoList in order to allow for multiple list
functionality. As a result some changes in how update is implemented,
but for the most part is the same.